Se creo EndPoint para registrar Usuario que incluye la creación de Usuario, CuentaFiduciaria y CuentaCripto en DB. También se implemento el patrón Repository en conjunto con el patrón UnitOfWork

+++ Entities/Historial.cs
@@ -24,7 +24,7 @@ namespace TrabajoFinalSofttek.Entities
 
         [Required]
         [Column("historial_monto")]
-        public long Monto { get; set; }
+        public decimal Monto { get; set; }
 
 
     }
